get-topic/requirements.txt
lxml==4.6.3
google-genai==1.23.0